% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_recorder_col_names.R
\name{change_recorder_col_names}
\alias{change_recorder_col_names}
\title{Convert column names from those output by Recorder to match QGIS ones}
\usage{
change_recorder_col_names(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Convert column names from those output by Recorder to match QGIS ones
}
