% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_and_N_calcs.R
\name{E_and_N_calcs}
\alias{E_and_N_calcs}
\title{Calculate eastings and northings based on grid ref}
\usage{
E_and_N_calcs(species_data)
}
\arguments{
\item{species_data}{Species data including a grid ref}
}
\value{

}
\description{
Calculate eastings and northings based on grid ref
}
\examples{


}
