% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_distance.R
\name{filter_by_distance}
\alias{filter_by_distance}
\title{Filter data by a given distance from a popint defined by eastings and northings}
\usage{
filter_by_distance(species_data, distance, grid_ref)
}
\arguments{
\item{species_data}{Species data}

\item{distance}{radius to remove beyond}

\item{grid_ref}{central grid ref}
}
\value{

}
\description{
Filter data by a given distance from a popint defined by eastings and northings
}
