% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_survey.R
\name{filter_by_survey}
\alias{filter_by_survey}
\title{Remove data from a couple of surveys}
\usage{
filter_by_survey(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{
filtered data
}
\description{
Remove data from a couple of surveys
}
