% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_GCN.R
\name{filter_GCN}
\alias{filter_GCN}
\title{Remove zero abundance Great crested newt records}
\usage{
filter_GCN(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Remove zero abundance Great crested newt records
}
