% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_designations.R
\name{fix_designations}
\alias{fix_designations}
\title{Fix designations as they're incorrect in Recorder - not sure if this is still needed}
\usage{
fix_designations(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Fix designations as they're incorrect in Recorder - not sure if this is still needed
}
