% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_check_data.R
\name{format_and_check_data}
\alias{format_and_check_data}
\title{Format the data for output and perform some checks}
\usage{
format_and_check_data(raw_data, OutputCols, newColNames, sensitivecheck)
}
\arguments{
\item{raw_data}{Set of species data}

\item{OutputCols}{List of the columns to be output}

\item{newColNames}{Labels for output columns}

\item{sensitivecheck}{Highlight sensitive species or not}
}
\value{

}
\description{
Format the data for output and perform some checks
}
\examples{
data <- format_and_check_data(raw_data,OutputCols,newColNames, 0)
}
