% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_check_EA_SLA_data.R
\name{format_and_check_EA_SLA_data}
\alias{format_and_check_EA_SLA_data}
\title{Format data for output for an SLA and perform some checks}
\usage{
format_and_check_EA_SLA_data(raw_data, EAOutputCols, newColNames)
}
\arguments{
\item{raw_data}{data to output}

\item{newColNames}{lables of output columns}

\item{OutputCols}{columns to output}
}
\value{

}
\description{
Format data for output for an SLA and perform some checks
}
\examples{
a <- c('','BirdsDir-A2.2')
b <- c('','')
c <- c('','')
d <- c('insect - butterfly','bird')
e <- c('Pieris rapae','Turdus merula')
f <- c('Small White','Blackbird')
g <- c('Mike Jeffries','Alnwick Wildlife Group')
h <- c('Morpeth Town','Morpeth Town')
i <- c('04/09/2010','04/09/2010')
j <- c('NZ18X','NZ18X')
k <- c('Butterfly Conservation','Alnwick Wildlife Group')
l <- c('Butterfly Conservation','Alnwick Wildlife Group')
m <- c('2 Count','1 Count')
n <- c('Considered Correct','Considered Correct')
o <- c(419000,419000)
p <- c(585000,585000)
q <- c(1000,1000)
r <- c('layer','layer')
s <- c('path','path')

df <- data.frame(a,a,b,c,d,d,e,f,a,g,h,h,i,j,k,l,m,n,o,p,q,q,r,s,b,b,b)
names(df) <- c('All.Design','AllUKLeg', 'Wildlife..', 'Wildlife_1', 'broad', 'Taxon.grou', 'Taxon.Lati', 'Taxon.Comm', 'Obs.Commen','Sample.Rec', 'Sample.Loc', 'Sample.L_1', 'Sample.Dat', 'Sample.Spa', 'Survey.Run', 'Survey.Nam', 'Obs.Abunda', 'Determinat', 'Central_eastings', 'Central_northings', 'Buffer','Precision', 'layer', 'path','NELBAP', 'ImportDate', 'Info')
OutputCols <- c("All.Design","AllUKLeg", "Wildlife..","Wildlife_1", "broad", "Taxon.grou", "Taxon.Lati","Obs.Abunda", "Determinat", "Obs.Commen","Sample.Rec","Sample.Loc",  "Sample.Dat","Sample.Spa",    "Central_eastings","Central_northings","Buffer","Precision","Survey.Run","Survey.Nam","NELBAP","Taxon.Comm","ImportDate","Info")
newColNames <- c("Designations","All UK Legally Protected (W&C Act, Badgers and Cons Regs) - Taxon Designated","Wildlife & Countryside Act Sch 9 Part 2","Wildlife & Countryside Act Sch 9 Part 1","Broad Group","Taxon group","Latin Name","Abundances","Determination Type","Comments","Recorder","Location Name","Date",  "Grid Reference","Cent_East","Cent_North","Buffer","Precision",             "Survey Run By","Survey Name",                     "North East LBAP - Short Names","Common Name","Obs Entry Date","Additional Information") # Three empty cols
data <- format_and_check_SLA_data(df,OutputCols,newColNames)
}
