% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_check_SLA_data.R
\name{format_and_check_SLA_data}
\alias{format_and_check_SLA_data}
\title{Format data for output for an SLA and perform some checks}
\usage{
format_and_check_SLA_data(raw_data, OutputCols, newColNames)
}
\arguments{
\item{newColNames}{}
}
\value{

}
\description{
Format data for output for an SLA and perform some checks
}
