% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DR_Excel_output.R
\name{format_DR_Excel_output}
\alias{format_DR_Excel_output}
\title{Format output data for a data request and output to an Excel spreadsheet}
\usage{
format_DR_Excel_output(
  XL_wb,
  sheet_name,
  outputdata,
  sensitivecheck,
  del_dist,
  DR_config,
  sheet_number
)
}
\arguments{
\item{XL_wb}{Excel workbook}

\item{sheet_name}{Label for Excel sheet}

\item{outputdata}{data set to be output}

\item{sensitivecheck}{Should sensitive species be highlighted?}

\item{del_dist}{Should the output be filtered by distance}

\item{DR_config}{list of config values}

\item{sheet_number}{Excel sheet number}
}
\value{

}
\description{
Format output data for a data request and output to an Excel spreadsheet
}
