% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SLA_Excel_output.R
\name{format_SLA_Excel_output}
\alias{format_SLA_Excel_output}
\title{format_SLA_Excel_output}
\usage{
format_SLA_Excel_output(XL_wb, sheet_name, outputdata, SLA_config, SLA_type)
}
\arguments{
\item{XL_wb}{Excel workbook}

\item{sheet_name}{name to give the worksheet}

\item{outputdata}{data to output}

\item{SLA_config}{list of config values}

\item{SLA_type}{whether it's a local authority SLA or NNPA/AONB}
}
\value{

}
\description{
format_SLA_Excel_output
}
