% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDates.R
\name{formatDates}
\alias{formatDates}
\title{formats Dates as far as possible given the variety that may be presented}
\usage{
formatDates(dateCol)
}
\arguments{
\item{dateCol}{Column containing date values}
}
\value{

}
\description{
formats Dates as far as possible given the variety that may be presented
}
