% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDurhamConfig.R
\name{getDurhamConfig}
\alias{getDurhamConfig}
\title{Setup config to split data for the Durham SLA}
\usage{
getDurhamConfig(bat_sp, plant_sp, Other_LA_Cols)
}
\arguments{
\item{Other_LA_Cols}{}
}
\value{

}
\description{
Setup config to split data for the Durham SLA
}
