% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorthumberlandConfig.R
\name{getNorthumberlandConfig}
\alias{getNorthumberlandConfig}
\title{Setup config to split data for the Northumberland SLA}
\usage{
getNorthumberlandConfig(plant_sp, Other_LA_Cols, SLA_split_config)
}
\arguments{
\item{SLA_split_config}{}
}
\value{

}
\description{
Setup config to split data for the Northumberland SLA
}
