% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitterConfig.R
\name{getSplitterConfig}
\alias{getSplitterConfig}
\title{Setup the config to split data depending on the local authority}
\usage{
getSplitterConfig(LA, SLA_Split_config)
}
\arguments{
\item{LA}{which local authority}

\item{SLA_Split_config}{config values}
}
\value{

}
\description{
Setup the config to split data depending on the local authority
}
\examples{
 config_keys <- c("bat_sp","plant_sp")
 bat_sp <- c("Chiroptera","Rhinolophidae","Vespertilionidae","Myotis","Barbastella","Eptesicus","Nyctalus","Pipistrellus","Plecotus","Vespertilio","Lasiurus","Tadarida","Hypsugo")
 plant_sp <- c("flowering plant" ,"clubmoss" ,"conifer" ,"fern" ,"fungus" ,"ginkgo" ,"horsetail" ,"lichen" ,"liverwort" ,"moss" ,"slime mould" ,"stonewort" ,"alga" ,"hornwort" ,"quillwort")
 config_values <- list(bat_sp,plant_sp)
 config_pairs <- list()                     # Create empty list
 for(i in 1:length(config_keys)) {              # Add key/value pairs in for-loop
   config_pairs[config_keys[i]] <- config_values[i]
 }
config <- getSplitterConfig(1,config_pairs)
}
