% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunderlandConfig.R
\name{getSunderlandConfig}
\alias{getSunderlandConfig}
\title{Setup config to split data for the Sunderland SLA}
\usage{
getSunderlandConfig(bat_sp, plant_sp)
}
\arguments{
\item{plant_sp}{}
}
\value{

}
\description{
Setup config to split data for the Sunderland SLA
}
