% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_location_cols.R
\name{merge_location_cols}
\alias{merge_location_cols}
\title{Merge the two locations columns into one}
\usage{
merge_location_cols(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Merge the two locations columns into one
}
