% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_manual_desigs.R
\name{remove_manual_desigs}
\alias{remove_manual_desigs}
\title{Filter out designations we used to remove manually}
\usage{
remove_manual_desigs(pandn_data, config_vals)
}
\arguments{
\item{config_vals}{list of config values}
}
\value{

}
\description{
Filter out designations we used to remove manually
}
