% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_optional_desigs.R
\name{remove_optional_desigs}
\alias{remove_optional_desigs}
\title{Remove the optional designations selected by the user}
\usage{
remove_optional_desigs(pandn_data, optional_desigs, config_vals)
}
\arguments{
\item{pandn_data}{data to process}

\item{optional_desigs}{Designations selected by the user}

\item{config_vals}{List of config values}
}
\value{

}
\description{
Remove the optional designations selected by the user
}
\examples{
data <- remove_optional_desigs(pandn_data,optional_desigs,config_vals)
}
