% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_protected_flag.R
\name{set_protected_flag}
\alias{set_protected_flag}
\title{Set the protected flag for appropriate species}
\usage{
set_protected_flag(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Set the protected flag for appropriate species
}
\examples{
a <- c('','')
d <- c('insect - butterfly','bird')
e <- c('Pieris rapae','Turdus merula')
df <- data.frame(d,e,a)
names(df) <- c('Taxon.grou', 'Taxon.Lati','AllUKLeg')

data <- set_protected_flag(df)
}
