% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_bfly_species.R
\name{setup_bfly_species}
\alias{setup_bfly_species}
\title{Setup list of butterfly species}
\usage{
setup_bfly_species()
}
\value{

}
\description{
Setup list of butterfly species
}
\examples{
butterflies <- setup_bfly_species()
}
