% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_bto_breeding_codes.R
\name{setup_bto_breeding_codes}
\alias{setup_bto_breeding_codes}
\title{Read list of BTO breeding codes from a CSV file}
\usage{
setup_bto_breeding_codes()
}
\value{

}
\description{
Read list of BTO breeding codes from a CSV file
}
\examples{
breeding_codes <- setup_bto_breeding_codes()
}
