% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_desig_config_values.R
\name{setup_desig_config_values}
\alias{setup_desig_config_values}
\title{Setup a list of config keys and values for designations that need correcting}
\usage{
setup_desig_config_values()
}
\value{
list
}
\description{
Setup a list of config keys and values for designations that need correcting
}
\examples{
config_pairs <- setup_desig_config_values()
}
