% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_DR_config_values.R
\name{setup_DR_config_values}
\alias{setup_DR_config_values}
\title{Setup a list of config keys and values for data requests}
\usage{
setup_DR_config_values()
}
\value{
list
}
\description{
Setup a list of config keys and values for data requests
}
\examples{
DRconfig <- setup_DR_config_values()
}
