% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_emals_to_ignore.R
\name{setup_emails_to_ignore}
\alias{setup_emails_to_ignore}
\title{Real list of email addresses to ignore from a csv file}
\usage{
setup_emails_to_ignore()
}
\value{

}
\description{
Real list of email addresses to ignore from a csv file
}
\examples{
emls <- setup_emails_to_ignore()
}
