% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fish_labels.R
\name{setup_fish_labels}
\alias{setup_fish_labels}
\title{Setup list of labels to use for fish taxon}
\usage{
setup_fish_labels()
}
\value{

}
\description{
Setup list of labels to use for fish taxon
}
\examples{
fish <- setup_fish_labels()
}
