% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_input_config_values.R
\name{setup_input_config_values}
\alias{setup_input_config_values}
\title{Setup a list of config keys and values for checking input data}
\usage{
setup_input_config_values()
}
\value{
list
}
\description{
Setup a list of config keys and values for checking input data
}
\examples{
input_config <- setup_input_config_values()
}
