% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_invasives_config.R
\name{setup_invasives_config}
\alias{setup_invasives_config}
\title{Read list of invasive species to flag on input}
\usage{
setup_invasives_config()
}
\value{

}
\description{
Read list of invasive species to flag on input
}
\examples{
values <- setup_invasives_config()
}
