% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_locs_to_ignore.R
\name{setup_locs_to_ignore}
\alias{setup_locs_to_ignore}
\title{Read list of locations to ignore from a CSV file}
\usage{
setup_locs_to_ignore()
}
\value{

}
\description{
Read list of locations to ignore from a CSV file
}
