% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_locs_to_replace.R
\name{setup_locs_to_replace}
\alias{setup_locs_to_replace}
\title{Read list of locations to replace from a CSV file}
\usage{
setup_locs_to_replace()
}
\value{

}
\description{
Read list of locations to replace from a CSV file
}
