% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_profanity_config.R
\name{setup_profanity_config}
\alias{setup_profanity_config}
\title{Read list of words for profanity check from a CSV file}
\usage{
setup_profanity_config()
}
\value{

}
\description{
Read list of words for profanity check from a CSV file
}
\examples{
words <- setup_profanity_config()
}
