% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_prot_config.R
\name{setup_prot_config}
\alias{setup_prot_config}
\title{Setup config for protected species}
\usage{
setup_prot_config()
}
\value{

}
\description{
Setup config for protected species
}
\examples{
config <- setup_prot_config
}
