% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_recorders_to_ignore.R
\name{setup_recorders_to_ignore}
\alias{setup_recorders_to_ignore}
\title{Real list of recorders to ignore from a csv file}
\usage{
setup_recorders_to_ignore()
}
\value{

}
\description{
Real list of recorders to ignore from a csv file
}
\examples{
recs <- setup_recorders_to_ignore()
}
