% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scientific_config.R
\name{setup_scientific_config}
\alias{setup_scientific_config}
\title{Read list of scientific phrases to flag on input}
\usage{
setup_scientific_config()
}
\value{

}
\description{
Read list of scientific phrases to flag on input
}
\examples{
values <- setup_scientific_config()
}
