% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sensitive_config.R
\name{setup_sensitive_config}
\alias{setup_sensitive_config}
\title{Setup config for senstive species and taxon groups}
\usage{
setup_sensitive_config()
}
\value{

}
\description{
Setup config for senstive species and taxon groups
}
\examples{
config <- setup_sensitive_config()
}
