% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_SLA_config_values.R
\name{setup_SLA_config_values}
\alias{setup_SLA_config_values}
\title{Setup confid values for SLAs}
\usage{
setup_SLA_config_values()
}
\value{

}
\description{
Setup confid values for SLAs
}
\examples{
config <- setup_SLA_config_values()
}
