% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_species_config.R
\name{setup_species_config}
\alias{setup_species_config}
\title{Read list of species to flag on input}
\usage{
setup_species_config()
}
\value{

}
\description{
Read list of species to flag on input
}
\examples{
words <- setup_species_config()
}
