% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_speciesnames_to_ignore.R
\name{setup_speciesnames_to_ignore}
\alias{setup_speciesnames_to_ignore}
\title{Real list of species names to ignore from a csv file}
\usage{
setup_speciesnames_to_ignore()
}
\value{

}
\description{
Real list of species names to ignore from a csv file
}
\examples{
spcs <- setup_speciesnames_to_ignore()
}
