% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_splitter_config_values.R
\name{setup_splitter_config_values}
\alias{setup_splitter_config_values}
\title{Setup config for SLA splitter}
\usage{
setup_splitter_config_values()
}
\value{

}
\description{
Setup config for SLA splitter
}
\examples{
config <- setup_splitter_config_values ()
}
