% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Sun_ST_cols.R
\name{setup_Sun_ST_cols}
\alias{setup_Sun_ST_cols}
\title{Setup column names for Sunderland and South Tyneside SLAs}
\usage{
setup_Sun_ST_cols()
}
\value{

}
\description{
Setup column names for Sunderland and South Tyneside SLAs
}
\examples{
 cols <- setup_Sun_ST_cols()
}
