% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Sun_ST_split.R
\name{setup_Sun_ST_split}
\alias{setup_Sun_ST_split}
\title{Setup config for the Sunderland and South Tyneside SLA splitter}
\usage{
setup_Sun_ST_split(bat_sp, plant_sp)
}
\arguments{
\item{bat_sp}{list of bat species}

\item{plant_sp}{list of plant species}
}
\value{

}
\description{
Setup config for the Sunderland and South Tyneside SLA splitter
}
