% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_survey_config_values.R
\name{setup_survey_config_values}
\alias{setup_survey_config_values}
\title{Setup list of config keys and values for specific surveys}
\usage{
setup_survey_config_values()
}
\value{
list
}
\description{
Setup list of config keys and values for specific surveys
}
\examples{
config <- setup_survey_config_values()
}
