% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_to_sheets.R
\name{split_data_to_sheets}
\alias{split_data_to_sheets}
\title{Split data to sheets within Excel workbook depending on the config specified}
\usage{
split_data_to_sheets(SLA_split, XL_wb, SLA_data, allDataSheet, outputCols)
}
\arguments{
\item{SLA_split}{config for the split rules}

\item{XL_wb}{Excel workbook}

\item{SLA_data}{data}

\item{allDataSheet}{name of the sheet to copy all data to}

\item{outputCols}{columns to output}
}
\value{

}
\description{
Split data to sheets within Excel workbook depending on the config specified
}
\examples{
 split_data_to_sheets(SLA_split,XL_wb,SLA_data,allDataSheet,outputCols)
}
