% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_house_numbers.R
\name{check_house_numbers}
\alias{check_house_numbers}
\title{Checks for house numbers in location name column. Deals with them as far as possible}
\usage{
check_house_numbers(inputdata, locsToReplace, locsToIgnore)
}
\arguments{
\item{locsToReplace}{\itemize{
\item tibble of locations to replace
}}

\item{locsToIgnore}{\itemize{
\item tibble locations to ignore
}}
}
\value{

}
\description{
Checks for house numbers in location name column. Deals with them as far as possible
}
\examples{

}
