% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_columns.R
\name{check_required_columns}
\alias{check_required_columns}
\title{check_required_columns}
\usage{
check_required_columns(data_set)
}
\arguments{
\item{data_set}{}
}
\value{

}
\description{
check_required_columns
}
