% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_temp_XL_cols.R
\name{delete_temp_XL_cols}
\alias{delete_temp_XL_cols}
\title{Delete ten columns from the given column in the Excel worksheet}
\usage{
delete_temp_XL_cols(XL_wb, sheet_number, delCol, delRow)
}
\arguments{
\item{delCol}{start column for deletion}

\item{delRow}{last row to delete}
}
\value{

}
\description{
Delete ten columns from the given column in the Excel worksheet
}
