% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDates.R
\name{formatDates}
\alias{formatDates}
\title{formats Dates as far as possible given the variety that may be presented}
\usage{
formatDates(dateCol)
}
\arguments{
\item{dateCol}{Column containing date values}
}
\value{

}
\description{
formats Dates as far as possible given the variety that may be presented
}
\examples{
e <- c('Pieris rapae','Turdus merula')
f <- c('Small White','Blackbird')
g <- c('Mike Jeffries','Alnwick Wildlife Group')
h <- c('Morpeth Town','Morpeth Town')
i <- c('04/09/2010','04/09/2010')
df <- data.frame(e,f,g,h,h,i)
names(df) <- c('Taxon.Lati', 'Taxon.Comm', 'Sample.Rec', 'Sample.Loc', 'Sample.L_1', 'Sample.Dat')
df$'Sample.Dat'<-formatDates(df$'Sample.Dat')
}
