% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGatesheadConfig.R
\name{getGatesheadConfig}
\alias{getGatesheadConfig}
\title{Setup config to split data for the Gateshead SLA}
\usage{
getGatesheadConfig(bat_sp, plant_sp, Other_LA_Cols)
}
\arguments{
\item{Other_LA_Cols}{}
}
\value{

}
\description{
Setup config to split data for the Gateshead SLA
}
