% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorthumberlandConfig.R
\name{getNorthumberlandConfig}
\alias{getNorthumberlandConfig}
\title{Setup config to split data for the Northumberland SLA}
\usage{
getNorthumberlandConfig(
  plant_sp,
  Other_LA_Cols,
  pearl_mussel_select,
  protected_mammals,
  non_inverts,
  s41_mammals
)
}
\arguments{
\item{plant_sp}{List of plant species}

\item{Other_LA_Cols}{List of columns}

\item{SLA_Split_config}{config values}
}
\value{

}
\description{
Setup config to split data for the Northumberland SLA
}
\examples{
bat_sp <- c("Chiroptera","Rhinolophidae","Vespertilionidae")
plant_sp <- c("flowering plant" ,"clubmoss" ,"conifer")
Other_LA_Cols <- c("Taxon group","Latin Name","Abundances","Determination Type")
pearl_mussel_select <- "SLA_data$`Latin Name`=='Margaritifera (Margaritifera) margaritifera' | SLA_data$`Latin Name`=='Margaritifera margaritifera'"
protected_mammals <- c("Arvicola amphibius", "Arvicola", "Lutra", "Lutra lutra", "Meles", "Meles meles", "Sciurus vulgaris", "Martes martes", "Martes", "Muscardinus avellanarius","Chiroptera","Rhinolophidae","Vespertilionidae","Myotis","Barbastella","Eptesicus","Nyctalus","Pipistrellus","Plecotus","Vespertilio","Lasiurus","Tadarida","Hypsugo")
non_inverts <- c("alga", "amphibian", "bird", "chromist", "clubmoss", "conifer", "diatom", "fern", "flowering plant", "fungus", "ginkgo", "hornwort", "horsetail", "lancelet (Cephalochordata)", "lichen", "liverwort", "marine mammal", "moss", "quillwort", "reptile", "slime mould", "stonewort", "terrestrial mammal","fish,")
s41_mammals <- c("Erinaceus europaeus", "Lepus europaeus", "Mustela putorius", "Micromys minutus")
list <- getNorthumberlandConfig(plant_sp,Other_LA_Cols,pearl_mussel_select,protected_mammals,non_inverts,s41_mammals)

}
