% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorthumberlandConfig.R
\name{getNorthumberlandConfig}
\alias{getNorthumberlandConfig}
\title{Setup config to split data for the Northumberland SLA}
\usage{
getNorthumberlandConfig(plant_sp, Other_LA_Cols, SLA_split_config)
}
\arguments{
\item{plant_sp}{List of plant species}

\item{Other_LA_Cols}{List of columns}

\item{SLA_split_config}{config values}
}
\value{

}
\description{
Setup config to split data for the Northumberland SLA
}
\examples{
list <- getNorthumberlandConfig(plant_sp,Other_LA_Cols,SLA_split_config)
}
