% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNorthumberlandConfig.R
\name{getNorthumberlandConfig}
\alias{getNorthumberlandConfig}
\title{Setup config to split data for the Northumberland SLA}
\usage{
getNorthumberlandConfig(plant_sp, Other_LA_Cols, SLA_Split_config)
}
\arguments{
\item{plant_sp}{List of plant species}

\item{Other_LA_Cols}{List of columns}

\item{SLA_split_config}{config values}
}
\value{

}
\description{
Setup config to split data for the Northumberland SLA
}
\examples{
bat_sp <- c("Chiroptera","Rhinolophidae","Vespertilionidae")
plant_sp <- c("flowering plant" ,"clubmoss" ,"conifer")
Other_LA_Cols <- c("Taxon group","Latin Name","Abundances","Determination Type")
  config_keys <- c("short_desigs","bat_sp","plant_sp","pearl_mussel_select","protected_mammals","non_inverts","s41_mammals")
short_desigs <- c('Durham BAP' = 'DuBAP','Newcastle BAP' = 'NeBAP','North Tyneside BAP' = 'NTBAP','Northumberland BAP' = 'NoBAP','Northumberland NP BAP' = 'NNPBAP','Tees Valley BAP' = 'TVBAP','England NERC' = 'EngNERC','England_NERC' = 'EngNERC','Wales NERC' = 'WalNERC','Wales_NERC' = 'WalNERC','Scottish Biodiversity List' = 'ScotBio','Scottish_Biodiversity_List' = 'ScotBio','RedList_Global' = 'RL_Glob','RedList_GB' = 'RL_GB','WACA-Sch5_sect9.1(kill/injuring)' = 'WACA5_9.1(kill/inj)','WACA-Sch5_sect9.1(taking)' = 'WACA5_9.1(tak)','WACA-Sch5_sect9.2' = 'WACA5_9.2','WACA-Sch5Sect9.4A*' = 'WACA5_9.4A*','WACA-Sch5_sect9.4.a' = 'WACA5_9.4.a','WACA-Sch5_sect9.4b' = 'WACA5_9.4b','WACA-Sch5Sect9.4c' = 'WACA5_9.4c','WACA-Sch5_sect9.5a' = 'WACA5_9.5a','WACA-Sch5_sect9.5b' = 'WACA5_9.5b')
pearl_mussel_select <- "SLA_data$`Latin Name`=='Margaritifera (Margaritifera) margaritifera' | SLA_data$`Latin Name`=='Margaritifera margaritifera'"
protected_mammals <- c("Arvicola amphibius", "Arvicola", "Lutra", "Lutra lutra", "Meles", "Meles meles", "Sciurus vulgaris", "Martes martes", "Martes", "Muscardinus avellanarius","Chiroptera","Rhinolophidae","Vespertilionidae","Myotis","Barbastella","Eptesicus","Nyctalus","Pipistrellus","Plecotus","Vespertilio","Lasiurus","Tadarida","Hypsugo")
non_inverts <- c("alga", "amphibian", "bird", "chromist", "clubmoss", "conifer", "diatom", "fern", "flowering plant", "fungus", "ginkgo", "hornwort", "horsetail", "lancelet (Cephalochordata)", "lichen", "liverwort", "marine mammal", "moss", "quillwort", "reptile", "slime mould", "stonewort", "terrestrial mammal","fish,")
s41_mammals <- c("Erinaceus europaeus", "Lepus europaeus", "Mustela putorius", "Micromys minutus")
config_values <- list(short_desigs,bat_sp,plant_sp,pearl_mussel_select,protected_mammals,non_inverts,s41_mammals)
config_pairs <- list()                     # Create empty list
for(i in 1:length(config_keys)) {              # Add key/value pairs in for-loop
 config_pairs[config_keys[i]] <- config_values[i]
}
list <- getNorthumberlandConfig(plant_sp,Other_LA_Cols,config_pairs)
}
