% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSouthTynesideConfig.R
\name{getSouthTynesideConfig}
\alias{getSouthTynesideConfig}
\title{Setup config to split data for the South Tyneside SLA (same as Sunderland)}
\usage{
getSouthTynesideConfig(bat_sp, plant_sp)
}
\arguments{
\item{bat_sp}{List of bat species}

\item{plant_sp}{List of plant species}
}
\value{

}
\description{
Setup config to split data for the South Tyneside SLA (same as Sunderland)
}
\examples{
bat_sp <- c("Chiroptera","Rhinolophidae","Vespertilionidae")
plant_sp <- c("flowering plant" ,"clubmoss" ,"conifer")
Other_LA_Cols <- c("Taxon group","Latin Name","Abundances","Determination Type")
list <- getSouthTynesideConfig(bat_sp,plant_sp)
}
