% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitterConfig.R
\name{getSplitterConfig}
\alias{getSplitterConfig}
\title{Setup the config to split data depending on the local authority}
\usage{
getSplitterConfig(LA, SLA_Split_config)
}
\arguments{
\item{LA}{which local authority}
}
\value{

}
\description{
Setup the config to split data depending on the local authority
}
