% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunderlandConfig.R
\name{getSunderlandConfig}
\alias{getSunderlandConfig}
\title{Setup config to split data for the Sunderland SLA}
\usage{
getSunderlandConfig(bat_sp, plant_sp)
}
\arguments{
\item{bat_sp}{List of bat species}

\item{plant_sp}{List of plant species}
}
\value{

}
\description{
Setup config to split data for the Sunderland SLA
}
\examples{
bat_sp <- c("Chiroptera","Rhinolophidae","Vespertilionidae")
plant_sp <- c("flowering plant" ,"clubmoss" ,"conifer")
list <- getSunderlandConfig(bat_sp,plant_sp)
}
