% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_location_cols.R
\name{merge_location_cols}
\alias{merge_location_cols}
\title{Merge the two locations columns into one}
\usage{
merge_location_cols(raw_data)
}
\arguments{
\item{raw_data}{data frame of raw data}
}
\value{

}
\description{
Merge the two locations columns into one
}
\examples{
h <- c('Morpeth Town','Morpeth Town')
df <- data.frame(h,h)
names(df) <- c('Sample.Loc', 'Sample.L_1')
data <- merge_location_cols(df)
}
