% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pandn_data.R
\name{process_pandn_data}
\alias{process_pandn_data}
\title{Filter data to just protected & notable species}
\usage{
process_pandn_data(pandn_data, DR_config, optional_desigs)
}
\arguments{
\item{pandn_data}{data to process}

\item{DR_config}{list of config values}

\item{optional_desigs}{optional designations selectd by the user}
}
\value{

}
\description{
Filter data to just protected & notable species
}
\examples{
pandn_data <- process_pandn_data(pandn_data,DR_config,optional_desigs)
}
