% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename colums}
\usage{
rename_columns(data_set, data_format)
}
\arguments{
\item{data_set}{Set of species data}

\item{data_format}{Source format of the data}
}
\value{

}
\description{
Rename colums
}
