% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_protected_flag.R
\name{set_protected_flag}
\alias{set_protected_flag}
\title{Set the protected flag for appropriate species}
\usage{
set_protected_flag(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{

}
\description{
Set the protected flag for appropriate species
}
\examples{
data <- set_protected_flag(raw_data)
}
